requirements.txt
python-dotenv
pyyaml
duckdb
psycopg2-binary
crowdstrike-falconpy
pytenable
requests
pandas
Jinja2
okta
duckdb
boto3
tabulate
sqlalchemy
zipp==3.19.1
pyarrow #==14.0.1
anyio==4.4.0
aiohttp==3.10.11
jwcrypto==1.5.6